Remove unnecessary unrelatedFunction to clean up the code
In this commit, the unrelatedFunction was removed because it was not contributing to the final functionality of the program. The function appeared to be added for debugging or testing purposes but was not necessary in the final version of the code. This cleanup improves the readability and maintainability of the code